refactor(dialogue_manager): extract apply.rs with pure free fns + tests
Step 2/9. Move apply_assessment, apply_extraction, apply_comprehension,
merge_concepts_into_turn into dialogue_manager/apply.rs along with
their 8 unit tests. Visibility tightens from pub(crate) to pub(super)
since the only callers are within dialogue_manager/. mod.rs imports
them via `use apply::{...}` so the rest of the file is unchanged.

mod.rs: 3572 → 3190 lines.
apply.rs: new, 427 lines (~150 fns + ~250 tests + doc).
